Feature - Closes #4
This commit adds a configuration script to the `Snipcart` API that gets loaded in right before it. It should safely parse and execute the code to add in the alternate cards, based on the Snipcart documentation, but I also wrapped it in a sanity check to alleviate any major failures.

From: https://github.com/DevLifts/devlifts-static/issues/4